+++ internal/executor/x/config.go
@@ -4,9 +4,9 @@ import (
 	_ "embed"
 
 	"github.com/kubeshop/botkube/internal/executor/x/getter"
-	"github.com/kubeshop/botkube/internal/plugin"
 	"github.com/kubeshop/botkube/pkg/api"
 	"github.com/kubeshop/botkube/pkg/config"
+	"github.com/kubeshop/botkube/pkg/pluginx"
 )
 
 var (
@@ -20,7 +20,7 @@ type Config struct {
 	Logger    config.Logger
 
 	// Fields not exposed to the user in the JSON schema
-	TmpDir plugin.TmpDir `yaml:"tmpDir"`
+	TmpDir pluginx.TmpDir `yaml:"tmpDir"`
 }
 
 // GetPluginDependencies returns exec plugin dependencies.
