+++ internal/source/kubernetes/recommendation/pod_no_latest_image_tag.go
@@ -11,6 +11,7 @@ import (
 	"github.com/kubeshop/botkube/internal/source/kubernetes/config"
 	"github.com/kubeshop/botkube/internal/source/kubernetes/event"
 	"github.com/kubeshop/botkube/internal/source/kubernetes/k8sutil"
+	"github.com/kubeshop/botkube/pkg/k8sx"
 )
 
 const podNoLatestImageTag = "PodNoLatestImageTag"
@@ -35,7 +36,7 @@ func (f *PodNoLatestImageTag) Do(_ context.Context, event event.Event) (Result, 
 	}
 
 	var pod coreV1.Pod
-	err := k8sutil.TransformIntoTypedObject(unstrObj, &pod)
+	err := k8sx.TransformIntoTypedObject(unstrObj, &pod)
 	if err != nil {
 		return Result{}, fmt.Errorf("while transforming object type %T into type: %T: %w", event.Object, pod, err)
 	}
