+++ pkg/pluginx/tmp_dir.go
@@ -1,18 +1,22 @@
-package plugin
+package pluginx
 
 import (
 	"os"
 	"path"
+
+	"github.com/kubeshop/botkube/internal/plugin"
 )
 
+// TmpDir represents temporary directory.
 type TmpDir string
 
+// Get returns temporary directory path.
 func (t TmpDir) Get() (string, bool) {
 	if t != "" {
 		return string(t), true
 	}
 
-	depDir := os.Getenv(DependencyDirEnvName)
+	depDir := os.Getenv(plugin.DependencyDirEnvName)
 	if depDir != "" {
 		return depDir, false
 	}
@@ -20,6 +24,7 @@ func (t TmpDir) Get() (string, bool) {
 	return path.Join(os.TempDir(), "bin"), true
 }
 
+// GetDirectory returns temporary directory.
 func (t TmpDir) GetDirectory() string {
 	dir, _ := t.Get()
 	return dir
